Clouds: static WebP in prod, live shader in dev
Production now serves a 3840×360 baked sprite (~13 KB) instead of
running the per-frame fragment shader. CSS scaleY(-1) handles the
header→footer mirror so one asset covers every placement. Dev still
lazy-loads the live WebGL shader for tuning.
